cloud_datawarehouse/requirements.txt
altair==4.2.2
annotated-types==0.7.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
configparser==7.2.0
cryptography==44.0.2
decorator==5.2.1
distro==1.9.0
entrypoints==0.4
executing==2.2.0
great-expectations==0.12.7
greenlet==3.1.1
idna==2.10
importlib_metadata==8.6.1
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_widgets==3.0.13
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mistune==3.1.3
monotonic==1.6
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.2
pandas==2.1.4
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.7
posthog==3.21.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==2.4.7
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.23.0
rpds-py==0.23.1
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
six==1.17.0
SQLAlchemy==1.4.54
sqlalchemy-redshift==0.8.14
stack-data==0.6.3
termcolor==2.5.0
toolz==1.0.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.25.11
wcwidth==0.2.13
widgetsnbextension==4.0.13
zipp==3.21.0
